Bug 1584936 - Ingest available measurement units
